+++ packages/Angular/Explorer/core-entity-forms/CHANGELOG.md
@@ -1,5 +1,53 @@
 # Change Log - @memberjunction/ng-core-entity-forms
 
+## 5.43.0
+
+### Patch Changes
+
+- Updated dependencies [40eb4e0]
+- Updated dependencies [9f6aa87]
+- Updated dependencies [9200b13]
+- Updated dependencies [a975e3d]
+- Updated dependencies [ad8d8f1]
+- Updated dependencies [a4cdfb0]
+- Updated dependencies [54183aa]
+  - @memberjunction/core@5.43.0
+  - @memberjunction/global@5.43.0
+  - @memberjunction/ai-core-plus@5.43.0
+  - @memberjunction/ai@5.43.0
+  - @memberjunction/core-entities@5.43.0
+  - @memberjunction/ng-base-forms@5.43.0
+  - @memberjunction/ng-ui-components@5.43.0
+  - @memberjunction/ng-tabstrip@5.43.0
+  - @memberjunction/ai-engine-base@5.43.0
+  - @memberjunction/actions-base@5.43.0
+  - @memberjunction/ng-base-application@5.43.0
+  - @memberjunction/ng-link-directives@5.43.0
+  - @memberjunction/ng-shared@5.43.0
+  - @memberjunction/ng-testing@5.43.0
+  - @memberjunction/ng-action-gallery@5.43.0
+  - @memberjunction/ng-actions@5.43.0
+  - @memberjunction/ng-agents@5.43.0
+  - @memberjunction/ng-ai-test-harness@5.43.0
+  - @memberjunction/ng-base-types@5.43.0
+  - @memberjunction/ng-code-editor@5.43.0
+  - @memberjunction/ng-deep-diff@5.43.0
+  - @memberjunction/ng-entity-relationship-diagram@5.43.0
+  - @memberjunction/ng-entity-viewer@5.43.0
+  - @memberjunction/ng-flow-editor@5.43.0
+  - @memberjunction/ng-join-grid@5.43.0
+  - @memberjunction/ng-list-management@5.43.0
+  - @memberjunction/ng-notifications@5.43.0
+  - @memberjunction/ng-record-process-studio@5.43.0
+  - @memberjunction/ng-search@5.43.0
+  - @memberjunction/ng-shared-generic@5.43.0
+  - @memberjunction/ng-timeline@5.43.0
+  - @memberjunction/ng-trees@5.43.0
+  - @memberjunction/ng-versions@5.43.0
+  - @memberjunction/graphql-dataprovider@5.43.0
+  - @memberjunction/templates-base-types@5.43.0
+  - @memberjunction/ng-markdown@5.43.0
+
 ## 5.42.0
 
 ### Patch Changes