+++ packages/Angular/Generic/deep-diff/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @memberjunction/ng-deep-diff
 
+## 5.43.0
+
+### Patch Changes
+
+- Updated dependencies [40eb4e0]
+- Updated dependencies [9f6aa87]
+- Updated dependencies [ad8d8f1]
+- Updated dependencies [a4cdfb0]
+  - @memberjunction/core@5.43.0
+  - @memberjunction/global@5.43.0
+
 ## 5.42.0
 
 ### Patch Changes